fix: reduce age number based on otolith counts
It did not appear that the otolith number reflected only unread otoliths but rather all otoliths collected that year. There were numerous instances where the number of aged fish = number of otoliths which pointed to this issue. Need to talk with CDFW next year about adjusting how the otoliths are reported.